refactor: remove duplication of function to generate expense data
